code clean up and renaming
Reduce one level of inheritence.
